Update mbedignore.py
Under Windows 10 the script would fail with an error similar to: 'OSError: [WinError 123] The filename, directory name, or volume label syntax is incorrect: 'C:\\Users\\Chris\\.platformio\\packages\\framework-mbed\\features\\frameworks\\utest\\*\\.mbedignore'

After running a debugger over the code, it appears that (in this case) the Path library does not operate the same way on Linux as it does on Windows. (This is a guess.)

This fix removes all calls to the Path library and replaces them with the equivalent `os.path` functions. For example, `Path(mbedignore_path).is_file()` becomes `os.path.isfile(mbedignore_path)`.